MIT1/backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL adapter
aiosqlite==0.19.0  # Async SQLite

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
aiohttp==3.9.1
httpx==0.25.2

# Data Processing
pandas==2.1.4
# Python 3.12-compatible NumPy
numpy==1.26.4
scikit-learn==1.3.2

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2
transformers==4.36.2
# Use a Python 3.12-compatible CPU build of PyTorch
torch==2.2.0+cpu

# API Integration
openai==1.3.8
arxiv==2.1.0
requests==2.31.0
aiohttp==3.9.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# Logging & Monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing HTTP clients

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Environment & Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# File Processing
python-magic==0.4.27
PyPDF2==3.0.1

# Caching
redis==5.0.1

# Web Scraping (for additional paper sources)
beautifulsoup4==4.12.2
lxml==4.9.3

# Scientific Computing
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0

# Additional Utilities
tqdm==4.66.1
click==8.1.7
rich==13.7.0

# API Documentation
## fastapi docs UIs not needed; FastAPI includes /docs and /redoc
## fastapi-swagger-ui==0.1.0
## fastapi-redoc==0.1.0

# Validation
email-validator==2.1.0

# Async Support
asyncio-mqtt==0.16.1

# Configuration Management
dynaconf==3.2.4

# Performance Monitoring
py-spy==0.3.14

# Database Migration
alembic==1.13.1

# CORS
# fastapi-cors==0.0.6

# Rate Limiting
slowapi==0.1.9

# File Upload
python-multipart==0.0.6

# JSON Schema
jsonschema==4.20.0

# Date/Time
python-dateutil==2.8.2

# Cryptography
cryptography==42.0.8

# WebSocket Support
websockets==12.0

# Task Queue
dramatiq==1.15.0

# Health Checks
healthcheck==1.3.3

# Metrics
prometheus-client==0.19.0

# Configuration
pydantic-settings==2.1.0

# Environment Variables
python-decouple==3.8

# API Versioning
fastapi-versioning==0.10.0

# Database Connection Pooling
sqlalchemy-pool==2.0.23

# Additional AI/ML Libraries
langchain==0.0.350
chromadb==0.4.18
faiss-cpu==1.7.4

# Text Processing
textstat==0.7.3
readability==0.3.1

# Citation Processing
pybtex==0.24.0
bibtexparser==1.4.0

# XML Processing
xmltodict==0.13.0

# Data Validation
cerberus==1.3.5

# Async Database
asyncpg==0.29.0  # Async PostgreSQL
aiomysql==0.2.0  # Async MySQL

# Additional HTTP Features
aiofiles==23.2.1

# Memory Profiling
memory-profiler==0.61.0

# Code Quality
bandit==1.7.5
safety==2.3.5

# Performance
## uvloop doesn't support Windows; skip for cross-platform simplicity
## uvloop==0.19.0  # Fast event loop for asyncio
